+++ db/import_column_family_job.cc
@@ -145,7 +145,8 @@ Status ImportColumnFamilyJob::Run() {
         file_metadata.smallest_seqno, file_metadata.largest_seqno, false,
         file_metadata.temperature, kInvalidBlobFileNumber, oldest_ancester_time,
         current_time, file_metadata.epoch_number, kUnknownFileChecksum,
-        kUnknownFileChecksumFuncName, f.unique_id, 0);
+        kUnknownFileChecksumFuncName, f.unique_id, 0,
+        f.table_properties.tail_start_offset);
     s = dummy_version_builder.Apply(&dummy_version_edit);
   }
   if (s.ok()) {