+++ db/version_edit.cc
@@ -238,6 +238,12 @@ bool VersionEdit::EncodeTo(std::string* dst) const {
                   f.compensated_range_deletion_size);
       PutLengthPrefixedSlice(dst, Slice(compensated_range_deletion_size));
     }
+    if (f.tail_start_offset) {
+      PutVarint32(dst, NewFileCustomTag::kTailStartOffset);
+      std::string varint_tail_start_offset;
+      PutVarint64(&varint_tail_start_offset, f.tail_start_offset);
+      PutLengthPrefixedSlice(dst, Slice(varint_tail_start_offset));
+    }
 
     TEST_SYNC_POINT_CALLBACK("VersionEdit::EncodeTo:NewFile4:CustomizeFields",
                              dst);
@@ -414,6 +420,11 @@ const char* VersionEdit::DecodeNewFile4From(Slice* input) {
         case kCompensatedRangeDeletionSize:
           if (!GetVarint64(&field, &f.compensated_range_deletion_size)) {
             return "Invalid compensated range deletion size";
+          }
+          break;
+        case kTailStartOffset:
+          if (!GetVarint64(&field, &f.tail_start_offset)) {
+            return "invalid tail start offset";
           }
           break;
         default:
@@ -851,6 +862,8 @@ std::string VersionEdit::DebugString(bool hex_key) const {
       InternalUniqueIdToExternal(&id);
       r.append(UniqueIdToHumanString(EncodeUniqueIdBytes(&id)));
     }
+    r.append(" tail start offset:");
+    AppendNumberTo(&r, f.tail_start_offset);
   }
 
   for (const auto& blob_file_addition : blob_file_additions_) {
@@ -966,6 +979,7 @@ std::string VersionEdit::DebugJSON(int edit_num, bool hex_key) const {
         // permanent
         jw << "Temperature" << static_cast<int>(f.temperature);
       }
+      jw << "TailStartOffset" << f.tail_start_offset;
       jw.EndArrayedObject();
     }