+++ table/block_based/block_based_table_builder.h
@@ -91,6 +91,10 @@ class BlockBasedTableBuilder : public TableBuilder {
   // is enabled.
   uint64_t EstimatedFileSize() const override;
 
+  // Get the offset where the "tail" part of a SST file starts. "Tail" refers to
+  // all blocks after data blocks till the end of the SST file.
+  uint64_t GetTailStartOffset() const override;
+
   bool NeedCompact() const override;
 
   // Get table properties