+++ table/block_based/block_based_table_factory.h
@@ -28,22 +28,6 @@ class BlockBasedTableBuilder;
 class RandomAccessFileReader;
 class WritableFileWriter;
 
-// A class used to track actual bytes written from the tail in the recent SST
-// file opens, and provide a suggestion for following open.
-class TailPrefetchStats {
- public:
-  void RecordEffectiveSize(size_t len);
-  // 0 indicates no information to determine.
-  size_t GetSuggestedPrefetchSize();
-
- private:
-  const static size_t kNumTracked = 32;
-  size_t records_[kNumTracked];
-  port::Mutex mutex_;
-  size_t next_ = 0;
-  size_t num_records_ = 0;
-};
-
 class BlockBasedTableFactory : public TableFactory {
  public:
   explicit BlockBasedTableFactory(
@@ -76,8 +60,6 @@ class BlockBasedTableFactory : public TableFactory {
 
   bool IsDeleteRangeSupported() const override { return true; }
 
-  TailPrefetchStats* tail_prefetch_stats() { return &tail_prefetch_stats_; }
-
  protected:
   const void* GetOptionsPtr(const std::string& name) const override;
   Status ParseOption(const ConfigOptions& config_options,
@@ -89,7 +71,6 @@ class BlockBasedTableFactory : public TableFactory {
  private:
   BlockBasedTableOptions table_options_;
   std::shared_ptr<CacheReservationManager> table_reader_cache_res_mgr_;
-  mutable TailPrefetchStats tail_prefetch_stats_;
 };
 
 extern const std::string kHashIndexPrefixesBlock;