+++ table/block_based/block_based_table_reader.cc
@@ -561,13 +561,13 @@ Status BlockBasedTable::Open(
     const InternalKeyComparator& internal_comparator,
     std::unique_ptr<RandomAccessFileReader>&& file, uint64_t file_size,
     uint8_t block_protection_bytes_per_key,
-    std::unique_ptr<TableReader>* table_reader,
+    std::unique_ptr<TableReader>* table_reader, uint64_t tail_start_offset,
+    bool contain_no_data_block,
     std::shared_ptr<CacheReservationManager> table_reader_cache_res_mgr,
     const std::shared_ptr<const SliceTransform>& prefix_extractor,
     const bool prefetch_index_and_filter_in_cache, const bool skip_filters,
     const int level, const bool immortal_table,
     const SequenceNumber largest_seqno, const bool force_direct_prefetch,
-    TailPrefetchStats* tail_prefetch_stats,
     BlockCacheTracer* const block_cache_tracer,
     size_t max_file_size_for_l0_meta_pin, const std::string& cur_db_session_id,
     uint64_t cur_file_num, UniqueId64x2 expected_unique_id) {
@@ -592,8 +592,9 @@ Status BlockBasedTable::Open(
 
   if (!ioptions.allow_mmap_reads) {
     s = PrefetchTail(ro, file.get(), file_size, force_direct_prefetch,
-                     tail_prefetch_stats, prefetch_all, preload_all,
-                     &prefetch_buffer, ioptions.stats);
+                     prefetch_all, preload_all, &prefetch_buffer,
+                     ioptions.stats, tail_start_offset, contain_no_data_block,
+                     ioptions.logger);
     // Return error in prefetch path to users.
     if (!s.ok()) {
       return s;
@@ -773,16 +774,6 @@ Status BlockBasedTable::Open(
       prefetch_all, table_options, level, file_size,
       max_file_size_for_l0_meta_pin, &lookup_context);
 
-  if (s.ok()) {
-    // Update tail prefetch stats
-    assert(prefetch_buffer.get() != nullptr);
-    if (tail_prefetch_stats != nullptr) {
-      assert(prefetch_buffer->min_offset_read() < file_size);
-      tail_prefetch_stats->RecordEffectiveSize(
-          static_cast<size_t>(file_size) - prefetch_buffer->min_offset_read());
-    }
-  }
-
   if (s.ok() && table_reader_cache_res_mgr) {
     std::size_t mem_usage = new_table->ApproximateMemoryUsage();
     s = table_reader_cache_res_mgr->MakeCacheReservation(
@@ -805,23 +796,32 @@ Status BlockBasedTable::Open(
 
 Status BlockBasedTable::PrefetchTail(
     const ReadOptions& ro, RandomAccessFileReader* file, uint64_t file_size,
-    bool force_direct_prefetch, TailPrefetchStats* tail_prefetch_stats,
-    const bool prefetch_all, const bool preload_all,
-    std::unique_ptr<FilePrefetchBuffer>* prefetch_buffer, Statistics* stats) {
+    bool force_direct_prefetch, const bool prefetch_all, const bool preload_all,
+    std::unique_ptr<FilePrefetchBuffer>* prefetch_buffer, Statistics* stats,
+    uint64_t tail_start_offset, bool contain_no_data_block,
+    Logger* const logger) {
+  assert(tail_start_offset <= file_size);
+
   size_t tail_prefetch_size = 0;
-  if (tail_prefetch_stats != nullptr) {
-    // Multiple threads may get a 0 (no history) when running in parallel,
-    // but it will get cleared after the first of them finishes.
-    tail_prefetch_size = tail_prefetch_stats->GetSuggestedPrefetchSize();
-  }
-  if (tail_prefetch_size == 0) {
-    // Before read footer, readahead backwards to prefetch data. Do more
-    // readahead if we're going to read index/filter.
+  if (tail_start_offset != 0) {
+    tail_prefetch_size = file_size - tail_start_offset;
+  } else if (contain_no_data_block) {
+    tail_prefetch_size = file_size - tail_start_offset;
+    ROCKS_LOG_WARN(logger,
+                   "Tail prefetch size %zu is calculated based on tail start "
+                   "offset being 0 and the "
+                   "file does not contain any data blocks",
+                   tail_prefetch_size);
+  } else {
     // TODO: This may incorrectly select small readahead in case partitioned
     // index/filter is enabled and top-level partition pinning is enabled.
-    // That's because we need to issue readahead before we read the properties,
+    // That's because we need to issue readahead before we read the
+    // properties,
     // at which point we don't yet know the index type.
     tail_prefetch_size = prefetch_all || preload_all ? 512 * 1024 : 4 * 1024;
+    ROCKS_LOG_WARN(logger,
+                   "Tail prefetch size %zu is calculated based on heuristics",
+                   tail_prefetch_size);
   }
   size_t prefetch_off;
   size_t prefetch_len;
@@ -1140,7 +1140,8 @@ Status BlockBasedTable::PrefetchIndexAndFilterBlocks(
   // are hence follow the configuration for pin and prefetch regardless of
   // the value of cache_index_and_filter_blocks
   if (prefetch_all || pin_partition) {
-    s = rep_->index_reader->CacheDependencies(ro, pin_partition);
+    s = rep_->index_reader->CacheDependencies(ro, pin_partition,
+                                              prefetch_buffer);
   }
   if (!s.ok()) {
     return s;
@@ -1164,7 +1165,7 @@ Status BlockBasedTable::PrefetchIndexAndFilterBlocks(
     if (filter) {
       // Refer to the comment above about paritioned indexes always being cached
       if (prefetch_all || pin_partition) {
-        s = filter->CacheDependencies(ro, pin_partition);
+        s = filter->CacheDependencies(ro, pin_partition, prefetch_buffer);
         if (!s.ok()) {
           return s;
         }
@@ -1984,8 +1985,8 @@ Status BlockBasedTable::ApproximateKeyAnchors(const ReadOptions& read_options,
   // `CacheDependencies()` brings all the blocks into cache using one I/O. That
   // way the full index scan usually finds the index data it is looking for in
   // cache rather than doing an I/O for each "dependency" (partition).
-  Status s =
-      rep_->index_reader->CacheDependencies(read_options, false /* pin */);
+  Status s = rep_->index_reader->CacheDependencies(
+      read_options, false /* pin */, nullptr /* prefetch_buffer */);
   if (!s.ok()) {
     return s;
   }