+++ table/block_fetcher_test.cc
@@ -268,7 +268,8 @@ class BlockFetcherTest : public testing::Test {
     ASSERT_NE(table_options, nullptr);
     ASSERT_OK(BlockBasedTable::Open(
         ro, ioptions, EnvOptions(), *table_options, comparator, std::move(file),
-        file_size, 0 /* block_protection_bytes_per_key */, &table_reader));
+        file_size, 0 /* block_protection_bytes_per_key */, &table_reader,
+        0 /* tail_start_offset */, false /* contain_no_data_block */));
 
     table->reset(reinterpret_cast<BlockBasedTable*>(table_reader.release()));
   }